backend/requirements-test.txt
flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
marshmallow-sqlalchemy==0.22.2
Flask-Cors==3.0.8
pytest==5.3.5
coverage==5.0.3
PyHamcrest==2.0.0